feature: Add additional product and packaging fields
In preparation for the upcoming digital product passport feature, some
new fields needed to be added to the product and the product materials.
Additionally, the packaging materials needed to be rewritten to use the
same material structure as the product materials.
For better UI/UX experience, the materials display in the details
component has been rewritten to show materials as cards instead of plain
text list.